made the Langchain based chat bot functionality
