+++ Backend/Controllers/payment.js
@@ -28,7 +28,8 @@ router.post("/orders", async (req, res) => {
         if (!order) return res.status(500).send("Some error occured");
 
         res.json(order);
-    } catch (error) {
+    } 
+    catch (error) {
         res.status(500).send(error);
     }
 });
