+++ Backend/Models/Joi Schema/JoiBlogSchema.js
@@ -3,7 +3,7 @@ const Joi = require('joi');
 const blogJoiSchema = Joi.object({
     title: Joi.string().required(),
     description: Joi.string().required(),
-    selectedCategory: Joi.string().valid('education', 'technology', 'travel', 'health').required(),
+    selectedCategory: Joi.string().valid('education', 'technology', 'travel', 'health', 'parenting', 'finance', 'photography', 'business/entrepreneurship', 'music', 'food', 'sports', 'lifestyle', 'history').required(),
     content: Joi.string().allow('').optional(),
     username: Joi.string().optional(),
     image: Joi.string().optional(),
