+++ Backend/package.json
@@ -12,6 +12,8 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "@langchain/core": "^0.2.10",
+    "@langchain/google-genai": "^0.0.21",
     "bcryptjs": "^2.4.3",
     "cookie-parser": "^1.4.6",
     "cookieparser": "^0.1.0",
@@ -21,6 +23,7 @@
     "express-rate-limit": "^7.3.1",
     "joi": "^17.13.1",
     "jsonwebtoken": "^9.0.2",
+    "langchain": "^0.2.7",
     "mongoose": "^8.2.4",
     "multer": "^1.4.5-lts.1",
     "node-cron": "^3.0.3",
