--- Bruno-docs/Blog Routes/DELETE.bru
@@ -1,11 +1,0 @@
-meta {
-  name: DELETE
-  type: http
-  seq: 4
-}
-
-delete {
-  url: http://localhost:1111/blog/delete/6661daf8abb9c87b94b24fbb
-  body: none
-  auth: none
-}
