--- Bruno-docs/Blog Routes/PATCH.bru
@@ -1,11 +1,0 @@
-meta {
-  name: PATCH
-  type: http
-  seq: 3
-}
-
-patch {
-  url: http://localhost:1111/blog/update/
-  body: none
-  auth: none
-}
