--- Bruno-docs/Blog Routes/POST.bru
@@ -1,11 +1,0 @@
-meta {
-  name: POST
-  type: http
-  seq: 2
-}
-
-post {
-  url: http://localhost:1111/blog/createPost
-  body: none
-  auth: none
-}
