--- Bruno-docs/Comment Routes/DELETE.bru
@@ -1,11 +1,0 @@
-meta {
-  name: DELETE
-  type: http
-  seq: 4
-}
-
-delete {
-  url: http://localhost:1111/review/deleteComment/6661dbf3abb9c87b94b24fc0
-  body: none
-  auth: none
-}
