--- Bruno-docs/Comment Routes/GET.bru
@@ -1,11 +1,0 @@
-meta {
-  name: GET
-  type: http
-  seq: 1
-}
-
-get {
-  url: http://localhost:1111/review/comments
-  body: none
-  auth: none
-}
