--- Bruno-docs/Comment Routes/PUT.bru
@@ -1,11 +1,0 @@
-meta {
-  name: PUT
-  type: http
-  seq: 3
-}
-
-put {
-  url: http://localhost:1111/review/updateComment/
-  body: none
-  auth: none
-}
