--- Bruno-docs/User Routes/GET User Profile.bru
@@ -1,11 +1,0 @@
-meta {
-  name: GET User Profile
-  type: http
-  seq: 2
-}
-
-get {
-  url: http://localhost:1111/user/profile
-  body: none
-  auth: none
-}
