--- Bruno-docs/User Routes/POST SignIn.bru
@@ -1,11 +1,0 @@
-meta {
-  name: POST SignIn
-  type: http
-  seq: 3
-}
-
-post {
-  url: http://localhost:1111/user/signIn
-  body: none
-  auth: none
-}
