--- Bruno-docs/User Routes/POST SignUp.bru
@@ -1,11 +1,0 @@
-meta {
-  name: POST SignUp
-  type: http
-  seq: 4
-}
-
-post {
-  url: http://localhost:1111/user/signUp
-  body: none
-  auth: none
-}
