--- Bruno-docs/bruno.json
@@ -1,9 +1,0 @@
-{
-  "version": "1",
-  "name": "Bruno-docs",
-  "type": "collection",
-  "ignore": [
-    "node_modules",
-    ".git"
-  ]
-}
