+++ Front-End/src/Css/AIChat.css
@@ -1,0 +1,20 @@
+/* AIChat.css */
+
+@keyframes typewriter {
+    from { width: 0 }
+    to { width: 100% }
+  }
+  
+  .typewriter {
+    overflow: hidden; /* Ensures the text is not visible outside the containing div */
+    white-space: nowrap; /* Keeps the text on a single line */
+    animation: typewriter 2s steps(40, end); /* Adjust animation duration and steps as needed */
+  }
+  
+  .typewriter-text {
+    border-right: 0.1em solid transparent; /* Creates the blinking cursor effect */
+    font-family: monospace; /* Ensures uniform character width */
+    margin: 0 auto; /* Centers the text within the div */
+    letter-spacing: 0.1em; /* Adjusts spacing between characters */
+  }
+
