+++ Front-End/src/Css/SignIn.css
@@ -82,7 +82,6 @@
     padding: 17px 17px 17px ;
     font-size: 1.2vw;
     box-sizing: border-box;
-    /* margin: 15px 0 15px 0; */
 }
 
 .inputFields:focus {
@@ -370,7 +369,6 @@
         border: none;
         box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
         border-radius: 7px;
-        /* padding: 14px 15px 14px; */
         font-size: 3vw;
         margin-bottom: 5vw;
     }
