+++ Front-End/src/Css/SignUp.css
@@ -40,12 +40,12 @@
 .right-cont-Up{
     border-radius: 0 40px 40px 0;
     align-items: center;
-    padding: 44px 20px 20px 20px;
+    padding: 60px 20px 20px 20px;
     height: 38vw;
     width: 42vw;
     display: flex;
     flex-direction: column;
-    justify-content: space-between;
+    justify-content: space-evenly;
 }
 
 .left-cont-Up h1{
@@ -284,7 +284,6 @@
         border: none;
         box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
         border-radius: 7px;
-        padding: 14px 15px 14px;
         font-size: 1.8vw;
     }
     
@@ -294,7 +293,7 @@
     
 }
 
-@media (min-width: 350px) and (max-width: 549px) {
+@media (max-width: 549px) {
     .bothUpContainers {
         display: flex;
         flex-direction: column;
@@ -304,8 +303,8 @@
     .left-cont-Up {
         flex-direction: column;
         border-radius: 30px 0 0 0;
-        width: 72vw;
-        height: 50vw;
+        width: 78vw;
+        height: 72vw;
         padding: 18px 18px 18px 18px;
     }
 
@@ -313,7 +312,7 @@
         border-radius: 0 0 30px 0;
         display: flex;
         flex-direction: column;
-        width: 72vw;;
+        width: 78vw;;
         box-sizing: border-box;
         padding: 18px 18px 18px 18px;
         height: fit-content;
@@ -335,7 +334,6 @@
         border: none;
         box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
         border-radius: 7px;
-        padding: 14px 15px 14px;
         font-size: 3vw;
     }
     
@@ -350,3 +348,4 @@
     }
     
 }
+
